Move Cache types to Caching namespace
Not a breaking change.  These types were implemented in this branch
