+++ Testing/VelaptorTests/Content/AtlasDataTests.cs
@@ -9,6 +9,7 @@ namespace VelaptorTests.Content
     using System.IO.Abstractions;
     using Moq;
     using Velaptor.Content;
+    using Velaptor.Content.Caching;
     using Velaptor.Content.Exceptions;
     using Velaptor.Graphics;
     using VelaptorTests.Helpers;
