+++ Velaptor/Content/AtlasData.cs
@@ -11,6 +11,7 @@ namespace Velaptor.Content
     using System.Diagnostics.CodeAnalysis;
     using System.IO.Abstractions;
     using System.Linq;
+    using Velaptor.Content.Caching;
     using Velaptor.Content.Exceptions;
     using Velaptor.Graphics;
