+++ Velaptor/Content/AtlasLoader.cs
@@ -9,6 +9,7 @@ namespace Velaptor.Content
     using System.Diagnostics.CodeAnalysis;
     using System.IO.Abstractions;
     using System.Linq;
+    using Velaptor.Content.Caching;
     using Velaptor.Content.Exceptions;
     using Velaptor.Content.Factories;
     using Velaptor.Factories;
