+++ Velaptor/Content/Caching/IDisposableItemCache.cs
@@ -2,7 +2,7 @@
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace Velaptor.Content
+namespace Velaptor.Content.Caching
 {
     using System;
