+++ Velaptor/Content/Caching/IItemCache.cs
@@ -2,7 +2,7 @@
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace Velaptor.Content
+namespace Velaptor.Content.Caching
 {
     using System.Diagnostics.CodeAnalysis;
