+++ Velaptor/Content/TextureLoader.cs
@@ -8,6 +8,7 @@ namespace Velaptor.Content
     using System;
     using System.Diagnostics.CodeAnalysis;
     using System.IO.Abstractions;
+    using Velaptor.Content.Caching;
     using Velaptor.Content.Exceptions;
     using Velaptor.Factories;
